+++ src/composites.rs
@@ -9,7 +9,6 @@ use std::fmt;
 use std::future::Future;
 use std::net::SocketAddr;
 use std::ops::{Deref, DerefMut};
-use std::task::{Context, Poll};
 
 /// Trait for generating a default "not found" response. Must be implemented on
 /// the `Response` associated type for `MakeService`s being combined in a
@@ -41,9 +40,9 @@ impl<'a> HasRemoteAddr for &'a Option<SocketAddr> {
     }
 }
 
-impl<'a> HasRemoteAddr for &'a hyper::server::conn::AddrStream {
+impl HasRemoteAddr for Option<SocketAddr> {
     fn remote_addr(&self) -> Option<SocketAddr> {
-        Some(hyper::server::conn::AddrStream::remote_addr(self))
+        *self
     }
 }
 
@@ -56,15 +55,10 @@ impl<'a> HasRemoteAddr for &'a tokio::net::UnixStream {
 
 /// Trait implemented by services which can be composited.
 ///
-/// Wraps tower_service::Service
+/// Wraps hyper::Service
 pub trait CompositedService<ReqBody, ResBody, Error> {
-    /// See tower_service::Service::poll_ready
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Error>>;
-    /// See tower_service::Service::call
-    fn call(
-        &mut self,
-        req: Request<ReqBody>,
-    ) -> BoxFuture<'static, Result<Response<ResBody>, Error>>;
+    /// See hyper::Service::call
+    fn call(&self, req: Request<ReqBody>) -> BoxFuture<'static, Result<Response<ResBody>, Error>>;
 }
 
 impl<T, ReqBody, ResBody, Error> CompositedService<ReqBody, ResBody, Error> for T
@@ -72,31 +66,23 @@ where
     T: Service<Request<ReqBody>, Response = Response<ResBody>, Error = Error>,
     T::Future: Send + 'static,
 {
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Error>> {
-        Service::poll_ready(self, cx)
-    }
-
-    fn call(
-        &mut self,
-        req: Request<ReqBody>,
-    ) -> BoxFuture<'static, Result<Response<ResBody>, Error>> {
+    fn call(&self, req: Request<ReqBody>) -> BoxFuture<'static, Result<Response<ResBody>, Error>> {
         Box::pin(Service::call(self, req))
     }
 }
 
-type FutureService<ReqBody, ResBody, Error, MakeError> = BoxFuture<
+/// Type alias for the future returned by a `MakeService`
+pub type FutureService<ReqBody, ResBody, Error, MakeError> = BoxFuture<
     'static,
     Result<Box<dyn CompositedService<ReqBody, ResBody, Error> + Send>, MakeError>,
 >;
 
 /// Trait implemented by make services which can be composited.
 ///
-/// Wraps tower_service::Service
+/// Wraps hyper::Service
 pub trait CompositedMakeService<Target, ReqBody, ResBody, Error, MakeError> {
-    /// See tower_service::Service::poll_ready
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), MakeError>>;
-    /// See tower_service::Service::call
-    fn call(&mut self, target: Target) -> FutureService<ReqBody, ResBody, Error, MakeError>;
+    /// See hyper::Service::call
+    fn call(&self, target: Target) -> FutureService<ReqBody, ResBody, Error, MakeError>;
 }
 
 impl<T, S, F, Target, ReqBody, ResBody, Error, MakeError>
@@ -107,11 +93,7 @@ where
     F: Future<Output = Result<S, MakeError>> + Send + 'static,
     S: CompositedService<ReqBody, ResBody, Error> + Send + 'static,
 {
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), MakeError>> {
-        Service::poll_ready(self, cx)
-    }
-
-    fn call(&mut self, target: Target) -> FutureService<ReqBody, ResBody, Error, MakeError> {
+    fn call(&self, target: Target) -> FutureService<ReqBody, ResBody, Error, MakeError> {
         Box::pin(Service::call(self, target).map(|r| match r {
             Ok(s) => {
                 let s: Box<dyn CompositedService<ReqBody, ResBody, Error> + Send> = Box::new(s);
@@ -178,10 +160,9 @@ where
     }
 }
 
-impl<ReqBody, ResBody, Error, MakeError, Connection> Service<Connection>
+impl<ReqBody, ResBody, Error, MakeError> Service<Option<SocketAddr>>
     for CompositeMakeService<Option<SocketAddr>, ReqBody, ResBody, Error, MakeError>
 where
-    Connection: HasRemoteAddr,
     ReqBody: 'static,
     ResBody: NotFound<ResBody> + 'static,
     MakeError: Send + 'static,
@@ -191,27 +172,11 @@ where
     type Response = CompositeService<ReqBody, ResBody, Error>;
     type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
 
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        for service in &mut self.0 {
-            match service.1.poll_ready(cx) {
-                Poll::Ready(Ok(_)) => {}
-                Poll::Ready(Err(e)) => {
-                    return Poll::Ready(Err(e));
-                }
-                Poll::Pending => {
-                    return Poll::Pending;
-                }
-            }
-        }
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, target: Connection) -> Self::Future {
+    fn call(&self, target: Option<SocketAddr>) -> Self::Future {
         let mut services = Vec::with_capacity(self.0.len());
-        let addr = target.remote_addr();
-        for (path, service) in &mut self.0 {
+        for (path, service) in &self.0 {
             let path: &'static str = path;
-            services.push(service.call(addr).map_ok(move |s| (path, s)));
+            services.push(service.call(target).map_ok(move |s| (path, s)));
         }
         Box::pin(futures::future::join_all(services).map(|results| {
             let services: Result<Vec<_>, MakeError> = results.into_iter().collect();
@@ -275,19 +240,8 @@ where
     type Response = Response<ResBody>;
     type Future = BoxFuture<'static, Result<Response<ResBody>, Error>>;
 
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        for service in &mut self.0 {
-            match service.1.poll_ready(cx) {
-                Poll::Ready(Ok(_)) => {}
-                Poll::Ready(Err(e)) => return Poll::Ready(Err(e)),
-                Poll::Pending => return Poll::Pending,
-            }
-        }
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, req: Request<ReqBody>) -> Self::Future {
-        for &mut (base_path, ref mut service) in &mut self.0 {
+    fn call(&self, req: Request<ReqBody>) -> Self::Future {
+        for &(base_path, ref service) in &self.0 {
             if req.uri().path().starts_with(base_path) {
                 return service.call(req);
             }